+++ neurox/mechanism.h
@@ -84,11 +84,11 @@ class Mechanism {
   class StateVars {
    public:
     StateVars();
-    StateVars(short count, short *offsets, short *dv_offsets);
+    StateVars(int count, int *offsets, int *dv_offsets);
     ~StateVars();
-    short count_;         ///> number of cvode state variables
-    short *var_offsets_;  ///>offset of state vars in ml->data
-    short *dv_offsets_;   ///> offset of dx/dV for state vars
+    int count_;         ///> number of cvode state variables
+    int *var_offsets_;  ///>offset of state vars in ml->data
+    int *dv_offsets_;   ///> offset of dx/dV for state vars
   } * state_vars_;
 
   enum ModFunctions {
